Add separate password layout in admin edit screen
